fix ide warnings

+++ Player/PlayerBase.h
@@ -34,8 +34,8 @@ class PlayerBase {
     void CheckSurroundingTiles();
 
 
-    float position_x_;
-    float position_y_;
+    double position_x_;
+    double position_y_;
     int current_tile_;
     int start_tile_;
     int velocity_ = 150;
